boton borrar actualizado
el programa ya no cae, pero hay bug al seleccionar algo que sea de pedidos.

+++ Classes/delete/delete.cs
@@ -5,6 +5,7 @@ using System.Data;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using System.Windows.Controls;
 using static Microsoft.EntityFrameworkCore.DbLoggerCategory.Database;
 
@@ -12,13 +13,22 @@ namespace crud.Classes.delete
 {
     class Delete
     {
-        public void DeleteOrder(ListBox allOrdersList, MySqlConnection con)
+        public void DeleteOrder(ListBox Order, MySqlConnection con)
         {
-            string query = "DELETE FROM PEDIDO WHERE ID = @orderToDelete";
-            MySqlCommand command = new MySqlCommand(query, con);
+            MessageBoxResult deleteConfirm = MessageBox.Show($"¿Estas seguro de que deseas borrar este dato? \n{Order.SelectedValue.ToString()}",
+                                             "Borrar datos en la tabla", MessageBoxButton.YesNo);
 
-            command.Parameters.AddWithValue("@orderToDelete", allOrdersList.SelectedValue);
-            command.ExecuteNonQuery();
+            if (deleteConfirm == MessageBoxResult.Yes)
+            {
+                string query = "DELETE FROM PEDIDO WHERE ID = @orderToDelete";
+                MySqlCommand command = new MySqlCommand(query, con);
+
+                command.Parameters.AddWithValue("@orderToDelete", Order.SelectedValue);
+                command.ExecuteNonQuery();
+
+                MessageBox.Show("Borrado con exito");
+            }
+            
         }
 
     }
